Revamped the code to reduce bloat and improve performance--only for the FCFS, SJF, SJFPreempt, and Priority algorithms. Further updates will be made to the Lottery algorithm.
