+++ C Programs/Process Scheduler/README.txt
@@ -1,5 +1,5 @@
 //Author: Derek Coleman
-//Project: Encryption Program
-//Last Update: 3/6/19
+//Project: Process Scheduler
+//Last Update: 3/24/23
 
 scheduler.c: This .c file contains both the main function and the auxiliary functions neeeded to schedule the processes.
