+++ C Programs/Process Scheduler/input5.txt
@@ -1,0 +1,5 @@
+1,3,6,10,9 - 10
+2,6,2,1,7 - 8
+3,2,8,6,1 - 2
+4,1,3,11,5 - 6
+5,5,4,8,3 - 4
